+++ extensions/chroma/environment/block.ts
@@ -1,54 +1,62 @@
 import { LookupMethod } from '../../../types/beatmap/shared/chroma.ts';
 import { Vector3 } from '../../../types/beatmap/shared/heck.ts';
 import { IChromaEnvironment } from '../../../types/beatmap/v3/chroma.ts';
+import { IChromaEnvironmentBlock, IChromaEnvironmentPlacement } from '../types/environment.ts';
 
-/** **IMPORTANT:** Manually adjust block to be 1x1x1 unity unit (1x1x1 scale does usually not work) */
+/** **IMPORTANT:** Manually adjust block to be exactly 1x1x1 unity unit (1x1x1 scale does usually not work) */
 export class EnvironmentBlock implements IChromaEnvironment {
     id: string;
     lookupMethod: LookupMethod;
-    track?: string;
-    duplicate?: number;
-    active?: boolean;
+    track?: never;
+    readonly duplicate = 1;
+    active?: never;
     scale: Vector3;
-    position?: Vector3;
-    rotation?: Vector3;
-    localPosition?: Vector3;
-    localRotation?: Vector3;
-    lightID?: number;
-    origin: Vector3;
-    protected constructor(data: IChromaEnvironment, origin: Vector3) {
+    position: Vector3;
+    rotation: Vector3;
+    localPosition?: never;
+    localRotation?: never;
+    lightID?: never;
+    anchor: Vector3;
+    protected constructor(data: IChromaEnvironmentBlock, anchor: Vector3) {
         this.id = data.id;
         this.lookupMethod = data.lookupMethod;
         this.track = data.track;
-        this.duplicate = 1;
         this.active = data.active;
-        this.scale = data.scale || [1, 1, 1];
-        this.position = data.position;
-        this.rotation = data.rotation;
+        this.scale = data.scale ?? [1, 1, 1];
+        this.position = data.position ?? [0, 0, 0];
+        this.rotation = data.rotation ?? [0, 0, 0];
         this.localPosition = data.localPosition;
         this.localRotation = data.localRotation;
         this.lightID = data.lightID;
-        this.origin = origin;
+        this.anchor = anchor;
     }
-    static create(data: IChromaEnvironment, origin: Vector3) {
-        return new this(data, origin);
+    static create(data: IChromaEnvironmentBlock, anchor: Vector3) {
+        return new this(data, anchor);
     }
 
-    place(pos: Vector3, scale: Vector3, rotation: Vector3, insertTo?: never): IChromaEnvironment;
-    place(pos: Vector3, scale: Vector3, rotation: Vector3, insertTo: IChromaEnvironment[]): void;
-    place(pos: Vector3, scale: Vector3, rotation: Vector3, insertTo?: IChromaEnvironment[]): IChromaEnvironment | void {
+    place(options: IChromaEnvironmentPlacement, insertTo?: never): IChromaEnvironment;
+    place(options: IChromaEnvironmentPlacement, insertTo: IChromaEnvironment[]): void;
+    place(
+        options: IChromaEnvironmentPlacement,
+        insertTo?: IChromaEnvironment[],
+    ): IChromaEnvironment | void {
+        const scale = options.scale ? (this.scale.map((s, i) => s * options.scale![i]) as Vector3) : this.scale;
+        const position = options.position
+            ? (options.position.map((p, i) => p + this.anchor[i]) as Vector3)
+            : this.position;
+        const rotation = options.rotation
+            ? (options.rotation.map((r, i) => r + this.rotation[i]) as Vector3)
+            : this.rotation;
         const data = {
             id: this.id,
             lookupMethod: this.lookupMethod,
             track: this.track,
             duplicate: this.duplicate,
             active: this.active,
-            scale: [scale[0] * this.scale[0], scale[1] * this.scale[1], scale[2] * this.scale[2]],
-            position: [pos[0] + this.origin[0], pos[1] + this.origin[1], pos[2] + this.origin[2]],
-            rotation: [rotation[0], rotation[1], rotation[2]],
-            localPosition: this.localPosition,
-            localRotation: this.localRotation,
-            lightID: this.lightID,
+            scale,
+            position,
+            rotation,
+            lightID: options.lightID,
         } as IChromaEnvironment;
         if (insertTo) {
             insertTo.push(data);