+++ extensions/chroma/types/mod.ts
@@ -1,2 +1,3 @@
 export * from './color.ts';
 export * from './options.ts';
+export * from './environment.ts';